Enemy.cs: Correctly initialied StateMaching by including it inside the _Ready function and calling the GetNode function that is included by Godot.
EnemyState.cs:
Added more boilerplate functions to EnemyState, that being Process and Physics.

chase.cs:
Added _Ready to chase state

EnemyStateMachine.cs:
Made ENemy static so that it can be called directly and be part of the class and copy be shared.

Added a check to currentState in the EnemyStateMaching in Process as there is a possibility it could return Null

Changed the initialization of staes from states = [] to
states = new List<EnemyState>();

instead of calling states[0] you can call directly the ENemyState static copy