Feature/enemy (#31)
* label deleted. test scne in dev!

* Re-add Prototype Enemy to Correct Branch

The prototype enemy slowly moves towards the mouse.

* Placeholder Logic For Enemy.cs

Enemy should now have a way to be "damaged" and "die"

* Prepare Enemy Scene for Probable Redo

Added an Area2D for target detection and renamed elements of the enemy to make more sense

* Broke down Enemy.cs

Created EnemyState.cs and EnemyMovement.cs. Also fixed some typos.

* Updated Enemy to add Data resource

* Updates to make Enemy Resource to Set Enemy Stats

* Added State Machine to Enemy

I do not know for sure if this will work.

* Updates to StateMachine

EnemyStateMachine produces an error that there are no child nodes.

* Enemy.cs:
Correctly initialied StateMaching by including it inside the _Ready function and calling the GetNode function that is included by Godot.

EnemyState.cs:
Added more boilerplate functions to EnemyState, that being Process and Physics.

chase.cs:
Added _Ready to chase state

EnemyStateMachine.cs:
Made ENemy static so that it can be called directly and be part of the class and copy be shared.

Added a check to currentState in the EnemyStateMaching in Process as there is a possibility it could return Null

Changed the initialization of staes from states = [] to
states = new List<EnemyState>();

instead of calling states[0] you can call directly the ENemyState static copy

* Updated States to fix some issues

Enemy can now chase the cursor at an appropriate speed that can be set.

---------

Co-authored-by: Cristian <cristianpmason@icloud.com>
Co-authored-by: Ill-Come-Up-With-The-Name-Later <adtmax5@gmail.com>