ADDED feature
Added day night cycle controller
Correctly used a sin function to imitate day/night and a color palette for the colors.
Connected to the Ui so it correctly changes the time, but still not connected to actual day just time so far.

Created event that passes day, hour, and min so anyone can call it if they need the day and/or time