Feature/iguana (#85)
* Started the Iguana

The iguana is still immobile.

* Iguana Chases Player

Iguana chase now follows the player.

* Attack State and Timer

Iguana now has an attack state and a timer that will count when the attack actually occurs. The timer does not currently have a trigger for when it ends.

* State Swapping

Iguana can now swap between roaming and chase when the player comes into range.

* Iguana HurtBox Included

Iguana can now check if a player is in the attack range. Tweaks made to iguana detection radius.

* Work started on the Timers

Work started on using timers with the iguana. Note: Last time I ran this and tested the timer state swapping, I crashed Godot.

* Fixed Crashing

Iguana no longer crashes when swapping states sometimes.

* Iguana Roaming Now Works

Iguana will pick a position every 0.5s to move to if it has gotten within 50 units of its previously chosen position. It will continue to move towards its chosen position if it is not within 50 units.

* Iguana Attack Started

Iguana can now "attack" and will be able to attack multiple times if the player remains in its attack range.

* Iguana Lunge

The iguana now sort of lunges at the player when it is close.

* Removed lunging

The lunging barely works. Removed it in favor of probably just having a tail swipe animation instead.

* Iguana Hitbox

Iguana now has a hitbox.

* Iguana Fixes

Removed unnecessary data variable and should have fixed issue where the iguana stops moving.

* Changed when timer timeout registers attack

Timer now connects to attack in the Ready function for IguanaAttack

* Changed timer for IguanaRoam

IguanaRoam timer also sets its timeout function in the Ready method instead of EnterState

---------

Co-authored-by: Cristian Padleski <117920618+CR1SI@users.noreply.github.com>