Feature/driveable car (#103)
* Car Work Begun

Added in the basic function of the car. The main missing piece is smoother movement for rotating the sprite. Also issues with decelerating not happening when you press "S." The movement code will likely need to be rewritten.

* Car Movement Improvements

Car movement is a little bit smoother. Reversing still seems to not work. Some issues with setting the car's angle. Also the car has a very limited turning amount for some reason.

* Functional Car

The car is now functionally completed. It rotates and moves correctly.