PG goal marker (#105)
* Start on Implementing the Parking Spot

The parking spot has been started. It has some collision issues and does not seem to be able to detect when the car has entered it.

* Fixes

Car now draws on top of the parking space. The slingshotting effect on the car motion may also be fixed.

* More Issues

The car exists and does not exist.

* Fixed The Car Paradox

The car now only exists. The car is now detected correctly when entering the parking space.

* Car slingshotting patch.

The car should no longer have weird "wind up" velocity when you hold down accelerate and then release it.

* Decelerate Update

Fixed the fact that I broke deceleration with the last commit.