First part done!
Refactored hurtbox and Hitbox code. Functions on contact as it should, actual interactions will be done when the player plays the attack animation which is when hitbox wil be active womp womp.
setup ALL of the current masks and layers in the project to correctly relfect where they stand in the world and so interactions are not buggy.
Fixed weird error message coming from interaction manager.
patched some enemies, inclduing book enemy and some other ways on their roaming/walking veloity functions to account for velocity.

Made sure everything has hutbox and hitbox and collision shapes accordingly.

made the resoruces for each current enemy in.