finished
Made sample Ui for main menu, loading and saving fully functional.

If player has no save it will create a new save and change the button later to continue instead of new

we will add a make new game feature later

currently when u start the game you are put in the parking garage scene!

fixed issue where player would spawn anywhere, now it checks if the scene we are on is one that needs the player, if it isn't it won't spawn him!