Feature/save&load respawn system (#127)
* started work on save load first, respawn after

* saving works, adding screen next

* ship work finishing it

* loading fixed. Starting to make ui now

* finished

Made sample Ui for main menu, loading and saving fully functional.

If player has no save it will create a new save and change the button later to continue instead of new

we will add a make new game feature later

currently when u start the game you are put in the parking garage scene!

fixed issue where player would spawn anywhere, now it checks if the scene we are on is one that needs the player, if it isn't it won't spawn him!

* small tweak to driveable car who is technically still player!

* fixed it so, it works with new saves.

still did not add deleting a save file to make a new one, will do that later not now

* finished fixing menus, and connecting everything with save/load and settings

added shader
fixed scene manager
added camera to car and menu screen
fixed some stuff with menus and decoupling of scenes